refactor arbiter into orchestrator
